requirement.txt
confluent-kafka
google-cloud-pubsub
tensorflow-gpu==2.3.2
opency-python